Restructure default converters
