+++ lib/honey_format/configuration.rb
@@ -2,18 +2,17 @@
 
 module HoneyFormat
   # Holds HoneyFormat configuration
-  # @attr_reader [#call] header_converter the configured header converter
-  # @attr_reader [#call] converter the configured value converter
-  # @attr_writer [#call] header_converter to use
-  # @attr_writer [#call] converter the value converter to use
   class Configuration
-    attr_accessor :converter
-    attr_reader :header_converter
-
     # Instantiate configuration
     def initialize
-      @converter = ValueConverter.new
-      @header_converter = @converter[:header_column]
+      @converter = nil
+      @header_converter = nil
+    end
+
+    # Returns the header converter
+    # @return [#call] header_converter the configured header converter
+    def header_converter
+      @header_converter ||= converter[:header_column]
     end
 
     # Set the header converter
@@ -22,10 +21,47 @@ module HoneyFormat
     # @return [#call] the header converter
     def header_converter=(converter)
       @header_converter = if converter.is_a?(Symbol)
-                            @converter[converter]
+                            self.converter[converter]
                           else
                             converter
                           end
+    end
+
+    # Returns the value converter
+    # @attr_reader [#call] converter the configured value converter
+    def converter
+      @converter ||= ValueConverter.new
+    end
+
+    # Default value converters
+    # @return [Hash] hash with default converters
+    def default_converters
+      @default_converters ||= {
+        # strict variants
+        decimal!: StrictConvertDecimal,
+        integer!: StrictConvertInteger,
+        date!: StrictConvertDate,
+        datetime!: StrictConvertDatetime,
+        symbol!: StrictConvertSymbol,
+        downcase!: StrictConvertDowncase,
+        upcase!: StrictConvertUpcase,
+        boolean!: StrictConvertBoolean,
+        # safe variants
+        decimal: ConvertDecimal,
+        decimal_or_zero: ConvertDecimalOrZero,
+        integer: ConvertInteger,
+        integer_or_zero: ConvertIntegerOrZero,
+        date: ConvertDate,
+        datetime: ConvertDatetime,
+        symbol: ConvertSymbol,
+        downcase: ConvertDowncase,
+        upcase: ConvertUpcase,
+        boolean: ConvertBoolean,
+        md5: ConvertMD5,
+        hex: ConvertHex,
+        nil: ConvertNil,
+        header_column: ConvertHeaderColumn
+      }.freeze
     end
   end
 end
